backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery==5.3.4",
    "cryptography==41.0.7",
    "django==5.0.1",
    "django-cors-headers==4.3.1",
    "django-environ==0.11.2",
    "django-filter==23.5",
    "django-redis==5.4.0",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular==0.27.0",
    "gunicorn==21.2.0",
    "pillow==10.1.0",
    "psycopg2-binary==2.9.9",
    "python-decouple==3.8",
    "python-jose==3.3.0",
    "redis==5.0.1",
    "requests==2.31.0",
    "sendgrid==6.11.0",
    "stripe==7.9.0",
    "twilio==8.11.1",
    "whitenoise==6.6.0",
]
